order keys recursively, hashCode keys, move all fns to utils
